enemy spawner works
haven't fully tested it tho. also it only spawns one enemy every every 10s or whatever you set the interval to be. because I suspect spawning a bunch in at once would probably clog the game quickly but this Can be changed

don't ask why it suddenly works
refactoring issue between fov and enemy classes I think
