+++ Assets/Prefabs/Enemy/Enemy.prefab
@@ -111,6 +111,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6085622505290389798}
+  - component: {fileID: 6782541688706776760}
   - component: {fileID: 3360919407585783280}
   - component: {fileID: 1988794748266474298}
   - component: {fileID: 2023791185400357427}
@@ -141,6 +142,20 @@ Transform:
   - {fileID: 6406880145139136140}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &6782541688706776760
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4571238477727134883}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 754e7e588231a074a87d90860f42f2c2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  smoothSpeed: 2.5
+  setToDestroy: 0
 --- !u!114 &3360919407585783280
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -157,9 +172,9 @@ MonoBehaviour:
   alertLevel: 0
   movementSpeed: 1
   _attackSpeedSetter: 0.2
-  aggroSpeed: 1
-  radius: 6.0174193
-  angle: 116
+  aggroSpeed: 30
+  radius: 15
+  angle: 130
   playerRef: {fileID: 0}
   targetMask:
     serializedVersion: 2
